+++ CHANGELOG.md
@@ -87,6 +87,7 @@
 
 ### Unreleased
 
+- generate error bindings for custom errors [#1549](https://github.com/gakonst/ethers-rs/pull/1549)
 - Support overloaded events
   [#1233](https://github.com/gakonst/ethers-rs/pull/1233)
 - Relax Clone requirements when Arc<Middleware> is used